Prepare dashboard for multi-tab view
