+++ frontend/src/components/card/ProblemCard.tsx
@@ -12,7 +12,7 @@ type ProblemCardProps = {
 const Content = styled.div`
   display: block;
   margin: 10px;
-  padding: 10px 30px;
+  padding: 15px 30px;
   border-radius: 12px;
   box-shadow: 0 2px 8px rgba(0, 0, 0, 0.12);
   background-color: ${({ theme }) => theme.colors.white};
