+++ frontend/src/components/layout/CircleBackground.tsx
@@ -4,8 +4,6 @@ import { ThemeConfig } from '../config/Theme';
 import { FloatingCircle, Coordinate } from '../special/FloatingCircle';
 import { MainContainer, DynamicWidthContainer } from '../core/Container';
 import { Header } from '../navigation/Header';
-import { useAppSelector } from '../../util/Hook';
-import MinimalLayout from './MinimalLayout';
 
 const Content = styled.div`
   position: relative;
@@ -21,7 +19,6 @@ type MyProps = {
 }
 
 function CircleBackgroundLayout({ children }: MyProps) {
-  const { firebaseUser } = useAppSelector((state) => state.account);
   const [mousePosition, setMousePosition] = useState<Coordinate>({ x: 0, y: 0 });
 
   const mouseMoveHandler = useCallback((e: MouseEvent) => {
@@ -31,11 +28,6 @@ function CircleBackgroundLayout({ children }: MyProps) {
   useEffect(() => {
     window.onmousemove = mouseMoveHandler;
   }, [mouseMoveHandler]);
-
-  // If logged in, the home page becomes the dashboard, which uses a minimal layout
-  if (firebaseUser) {
-    return <MinimalLayout>{children}</MinimalLayout>;
-  }
 
   return (
     <Content>
